+++ whale-line-iconify.el
@@ -52,7 +52,8 @@ This has no effect if icons cannot be enabled. See
     (window-no-other . (:name "low-vision" :face whale-line-shadow :parent window-status))
     (buffer-fallback . (:name "question-circle" :face whale-line-contrast :no-defaults t))
     (lsp . (:name "plug" :face whale-line-contrast))
-    (debug . (:name "bug" :face whale-line-urgent)))
+    (debug . (:name "bug" :face whale-line-urgent))
+    (client . (:name "server" :face 'whale-line-shadow)))
   "Icon specifications.
 
 This is an alist of (ICON-SYMBOL . SPECS). The ICON-SYMBOL is the